requirements.txt
streamlit
qdrant-client
sentence-transformers
transformers
torch
pillow
pytesseract
opencv-python
imagehash
numpy
pandas
pydantic
python-dotenv
tqdm
requests
beautifulsoup4
